Translating this thing to german, part 2.
Part 2: Translating all of it. I'll probably need a few more days to get
the semantic stuff done. Meaning: Some of this is replaced with their
corresponding english translations. Yet they also sometimes lack
context, so this needs to be done while working with the system, which
will probably need a bit more "getting hands dirty" with the shop
system.

fixes #526