Let ransack know about translations
Since globalize 5 won't store translated data on default tables we need
to tweak the table / attribute name to let it fetch the data on the
right place.

This is just a proof of concept on what we could do to
improve the spree / ransack / globalize love here. Not sure this will
work with associations yet or even if it will work at all in all
scenarios